# specs/001-personal-growth-planner/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Growth Planner API
  description: |
    RESTful API for the Personal Growth Planner application.
    Supports OKR management, learning path tracking, calendar scheduling, and progress monitoring.
  version: 1.0.0
  contact:
    name: Agiemme Planner Team
    email: support@agiemme-planner.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.agiemme-planner.com/v1
    description: Production server
...
paths:
...
  /okrs/key-results/{keyResultId}:
    get:
      tags: [OKRs]
      summary: Get key result by ID
      operationId: getKeyResultById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/KeyResultId'
      responses:
        '200':
          description: Key result details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyResult'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [OKRs]
      summary: Update key result
      operationId: updateKeyResult
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/KeyResultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateKeyResultRequest'
      responses:
        '200':
          description: Key result updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyResult'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [OKRs]
      summary: Delete key result
      operationId: deleteKeyResult
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/KeyResultId'
      responses:
        '204':
          description: Key result deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /okrs/key-results/{keyResultId}/progress:
    post:
      tags: [OKRs]
      summary: Record progress update for key result
      operationId: createProgressUpdate
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/KeyResultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProgressUpdateRequest'
      responses:
        '201':
          description: Progress update recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressUpdate'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [OKRs]
      summary: Get progress history for key result
      operationId: getProgressUpdates
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/KeyResultId'
        - name: startDate
          in: query
          description: Filter progress updates from this date
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Filter progress updates until this date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Progress update history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProgressUpdate'
                  meta:
                    $ref: '#/components/schemas/ListMetadata'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ============================================================================
  # Learning Path Endpoints
  # ============================================================================
...
  /progress/at-risk:
    get:
      tags: [Progress]
      summary: Get key results that are at risk (behind schedule)
      operationId: getAtRiskKeyResults
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of at-risk key results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/KeyResultWithObjective'
                  meta:
                    $ref: '#/components/schemas/ListMetadata'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ============================================================================
  # User Profile Endpoints
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  parameters:
...
    KeyResultId:
      name: keyResultId
      in: path
      required: true
      description: Unique key result identifier
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ObjectiveCategory:
      type: string
      enum:
        - career
        - health
        - relationships
        - skills
        - financial
        - personal
    ObjectiveStatus:
      type: string
      enum:
        - active
        - completed
        - archived
        - abandoned
    KeyResultStatus:
      type: string
      enum:
        - not_started
        - in_progress
        - completed
        - at_risk
...
    Objective:
      type: object
      required:
        - id
        - userId
        - title
        - category
        - targetDate
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 5
          maxLength: 200
          example: Advance my career in software engineering
        description:
          type: string
          nullable: true
        category:
          $ref: '#/components/schemas/ObjectiveCategory'
        targetDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/ObjectiveStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    KeyResult:
      type: object
      required:
        - id
        - objectiveId
        - description
        - targetValue
        - currentValue
        - unit
        - deadline
        - completionPercentage
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        objectiveId:
          type: string
          format: uuid
        description:
          type: string
          minLength: 10
          maxLength: 300
          example: Complete 3 certifications
        targetValue:
          type: number
          format: decimal
          minimum: 0
          exclusiveMinimum: true
          example: 3.00
        currentValue:
          type: number
          format: decimal
          minimum: 0
          example: 1.00
        unit:
          type: string
          minLength: 1
          maxLength: 50
          example: certifications
        deadline:
          type: string
          format: date
        completionPercentage:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
          description: Calculated (currentValue / targetValue) * 100
        status:
          $ref: '#/components/schemas/KeyResultStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    KeyResultWithObjective:
      allOf:
        - $ref: '#/components/schemas/KeyResult'
        - type: object
          required:
            - objective
          properties:
            objective:
              $ref: '#/components/schemas/Objective'
    ProgressUpdate:
      type: object
      required:
        - id
        - keyResultId
        - valueRecorded
        - recordedAt
      properties:
        id:
          type: string
          format: uuid
        keyResultId:
          type: string
          format: uuid
        valueRecorded:
          type: number
          format: decimal
          minimum: 0
        notes:
          type: string
          maxLength: 1000
          nullable: true
        recordedAt:
          type: string
          format: date-time
...
    UpdateKeyResultRequest:
      type: object
      properties:
        description:
          type: string
          minLength: 10
          maxLength: 300
        targetValue:
          type: number
          format: decimal
          minimum: 0
          exclusiveMinimum: true
        currentValue:
          type: number
          format: decimal
          minimum: 0
        unit:
          type: string
          minLength: 1
          maxLength: 50
        deadline:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/KeyResultStatus'
    CreateProgressUpdateRequest:
      type: object
      required:
        - valueRecorded
      properties:
        valueRecorded:
          type: number
          format: decimal
          minimum: 0
        notes:
          type: string
          maxLength: 1000
...
    ListMetadata:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        count:
          type: integer
          description: Number of items in current response
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
          properties:
            message:
              type: string
              description: Human-readable error message
            code:
              type: string
              description: Machine-readable error code
    ValidationErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
            - validationErrors
          properties:
            message:
              type: string
              example: Validation failed
            validationErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: Field name that failed validation
                  message:
                    type: string
                    description: Validation error message
  responses:
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequestError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'